+++ amethyst_ui/src/prefab.rs
@@ -337,6 +337,8 @@ pub struct UiImagePrefab(UiImageLoadPrefab);
 pub enum UiImageLoadPrefab {
     /// A textured image
     Texture(TexturePrefab),
+    /// A partial textured image
+    PartialTexture(TexturePrefab, f32, f32, f32, f32),
     /// Solid color image
     SolidColor(f32, f32, f32, f32),
 }
@@ -384,6 +386,12 @@ impl<'a> PrefabData<'a> for UiImageLoadPrefab {
             UiImageLoadPrefab::Texture(tex) => {
                 UiImage::Texture(tex.add_to_entity(entity, textures, entities, children)?)
             }
+            UiImageLoadPrefab::PartialTexture(tex, left, right, bottom, top) => {
+                UiImage::PartialTexture(
+                    tex.add_to_entity(entity, textures, entities, children)?,
+                    [*left, *right, *bottom, *top].into(),
+                )
+            }
             UiImageLoadPrefab::SolidColor(r, g, b, a) => UiImage::SolidColor([*r, *g, *b, *a]),
         };
         Ok(image)
@@ -396,6 +404,7 @@ impl<'a> PrefabData<'a> for UiImageLoadPrefab {
     ) -> Result<bool, Error> {
         match self {
             UiImageLoadPrefab::Texture(tex) => tex.load_sub_assets(progress, textures),
+            UiImageLoadPrefab::PartialTexture(tex, ..) => tex.load_sub_assets(progress, textures),
             UiImageLoadPrefab::SolidColor(..) => Ok(false),
         }
     }